Fix linking and conflicts between raknet and raylib
Client and Server were not tested before the last commit, and were not
linked correctly, so that has been fixed. The main problem was conflicts
raylib and some of the libraries used by raknet. To fix this, the
functions CloseWindow and HideCursor have had the prefix Ray added to
them, and raylib::Rectangle is used in multiple places instead of
::Rectangle, due to a function of the same name existing in wingdi.h.