Implement input processing, Change server update model
Previously, input was implemented as a hack, with it only being used by
the server, and clients having to wait for it to send the result. Now it
is (mostly) implemented properly, with clients using it to predict their
movement and the server using it as intended.

IMPORTANT: Input has not been fully implemented yet. The input buffer
used by clients to apply predicted movement after receving an update from
the server has not been implemented yet, and nor has the ring buffer. As
of now, when an update from the server is receved, it will move forward
using dead reckoning as before.

Server flow has been altered for client objects. Before, inputs would be
applied in the past and dead reckoning would be used to predict forward.
Now client objects are only updated and sent out to clients as input is
receved and processed.