Finish implementation of input buffer, Make changes to input
The problem with applying input on top of a server update was that a
pre-processed diff state was being used, when a new one should have been
made using the original input. In order to do this, input processing
functions cant use bit streams, so they have been altered to use an
Input struct defined in ClientObject.h and used bu both Client and
Server.

An additional change is the message ID SERVER_SET_CLIENT_ID has been
changed to SERVER_CREATE_STATIC_OBJECTS, and no longer sends the
clientID. Instead, the client ID is receved with the client object.