Fix collision resolution, Implement inertia tensor
Collision resolution has been fixed and now works as intended. As part
of this, colliders can be added to an object with its contructor.

Moment is now calculated using a collider, as as moment of inertia is
dependant on shape. It has also been changed from a scalar to a tensor,
using a matrix.