Make changes to allow client physics
In preperation of having clients perform collision detection and some
resolution, the server now sends an objects mass, elasticity and
collider to clients, GameObject::resolveCollision now has additional
params for if its being used on the server and if the other object
should be affected, and GameObjects now have two onCollision events for
use on the server and on clients.

The input buffer now holds the client objects state at the time of the
input, which is used by ClientObject to determine if the prediction was
accurate, in which case it will continue to use the predicted value
instead of re-applying all inputs.

Client factory methods now take a struct containing object data that is
defined by the system, including PhysicsState, Collider, mass, and
elasticity.