Add client side collision
Clients now predict collisions. Currently, this is only done in the
physics update, not when applying a server update, but it works well
enough as is for that to not be an issue.

Game objects on clients are no longer extrapolated to the current time,
instead being 1/2 latency behind the server at all times. This will be
addressed when lag compensation is added to the server.