Add box-box collision, Fix rotational forces
OBB to OBB collision detection has been implemented, though the need to
generate a contact point has requiered a large number of extra functions
to be added with it.

Collison resolution has been 'fixed' multiple times now, but every time
only part of the problem had been fixed. This time, it has been fixed again,
this time more core to the problem. It still has some problem with the
normal force not correcting rotation, but its much better than it was.

Aditional changes include making the camera perspective again, drawing
everything as a box with 3D rotation, a static object for the ground,
and a temporary gravity.