Clean code, Remove some debuging tools, Update comments
Cleaning includes some reformatting, organisation, etc. One noteworthy
change is position and rotation in StaticObject are now protected with
getters. Because CollisionSystem needs to set position directly, it has
been made a friend class.

Nearing the compleation of the system, much of the code used for debuging
is no longer nessesary. The code used to run the test demo is still in
place, however.

Function comments have been updated, with all functions having a description.