Remove test application, Fix build settings
All temparary function definitions have been removed, along with the main
functions. The system is done, and testing stuff can be removed.

The settings have been fixed to work for release as well as debug. It
should also work for x64, but needs to be tested. Server and Client now
build as .lib rather than .exe since it is distributed as a library.