Fix build settings
Upon linking the libs to a new project, it has been discovered that, for
some reason, raknet will not work with the runtime library setting used
for the release build, although it would work within this project. To
fix the problem, the setting has been changed for the release build.