Combine seperate onCollsion calls into one
There were seperate onCollsions functions for the server and client, with
the expectation that they would be used for different functionality. In
reality, any time this would be nessesary, they would be deffined within
both server and client projects, so they can interact with the server and
client classes, making the two functions useless.