Fix static object fragmentation, Fix client collision
When the server sends all static objects in a single packet, an issue
can occur if there is variation in latency. I am unsure of exactly what
happens, but packet fragmentation causes packets to become broken,
causing issues for clients. To fix this, static objects packets will be
kept under the MTU to prevent fragmentation by sending the packet and
starting a new one.

Clients would check for collision with their client object without
checking that it existed, which could cause an error when concidering
latency varience.