Fix bug with object destruction
It was possible for clients to receve a create and destroy call in the
wrong order and have a game object that does not exist on the server. To
fix this, clients will check if an object with the ID to destroy exists,
and if not, add it to a 'blacklist'. When creating an object, it will
check if the new ID is in the blacklist, and not create the object if
so.