Add frame indexing to inputs
When sending inputs from a client to the server, the time stamp was used,
causing multiple issues reguarding timming. Now that a frame index is
being used, the timming can be much more accurate.

There are currently some bugs present, with object creation not working,
calculated input time drifting on the server, and frame rate dependence.